+++ core/common/src/main/kotlin/com/github/jcornaz/collekt/ExpectedFacories.kt
@@ -1,5 +1,14 @@
 package com.github.jcornaz.collekt
 
 import com.github.jcornaz.collekt.api.PersistentListFactory
+import com.github.jcornaz.collekt.api.PersistentSetFactory
 
+/**
+ * Default [PersistentListFactory]
+ */
 public expect var DefaultListFactory: PersistentListFactory
+
+/**
+ * Default [PersistentListFactory]
+ */
+public expect var DefaultSetFactory: PersistentSetFactory