+++ core/jvm6/src/main/kotlin/com/github/jcornaz/collekt/ActualFactories.kt
@@ -1,6 +1,9 @@
 package com.github.jcornaz.collekt
 
 import com.github.jcornaz.collekt.api.PersistentListFactory
+import com.github.jcornaz.collekt.api.PersistentSetFactory
+import com.github.jcornaz.collekt.dexx.DexxSet
 import com.github.jcornaz.collekt.dexx.DexxVector
 
 public actual var DefaultListFactory: PersistentListFactory = DexxVector
+public actual var DefaultSetFactory: PersistentSetFactory = DexxSet